expose config logic and generalize web target to support electron
